requirements.txt
pandas
sqlalchemy
requests
psycopg2-binary
python-dotenv
matplotlib
scikit-learn
streamlit